KVM: arm64: Add support for IDSR exits to userspace
As per ARM DDI 0487E.a, section D12.3.2, there can be various
system registers that are IMPLEMENTATION DEFINED.

So far, KVM would inject an undefined instruction into the guest,
whenever access to an implementation defined system register (from
here on referred to as IDSR) was trapped. This makes sense, since a
general-purpose OS probably shouldn't rely on the existence of IDSRs.

This patch introduces an option to give userspace a chance to handle
these traps. This can be used to emulate specific architectures, and
virtualize operating systems that rely on the existence of specific
IDSRs.

Similarly to the recently introduced NISV exits, this is an ABI to
userspace, and comes with a matching new capability that allows the
configuration of the behavior.

This patch introduces KVM_CAP_ARM_IDSR_TO_USER, which can be used to
enable userspace exits due to access to IDSRs. Additionally, it
introduces a matching new exit reason, KVM_EXIT_ARM_IDSR, to report
those events to userspace. Userspace can choose to emulate the access
based on the architecture requirements, or refuse to emulate it and
let the kernel continue with the default behavior (injecting an
undefined instruction exception into the guest).

Signed-off-by: Lev Aronsky <aronsky@gmail.com>